# Improve performance of method haveAllNodes