--- a/pkg/scheduler/framework/plugins/dynamicresources/dynamicresources_test.go
+++ b/pkg/scheduler/framework/plugins/dynamicresources/dynamicresources_test.go
@@ -227,6 +227,7 @@ func BenchmarkHaveAllNodes(b *testing.B) {
 			test.nodeNames = append(test.nodeNames, "worker-"+strconv.Itoa(i))
 			test.nodes = append(test.nodes, &st.MakeNode().Name("worker-"+strconv.Itoa(i)).Node)
 		}
+		b.ResetTimer()
 		b.Run(test.name, func(b *testing.B) {
 			for i := 0; i < b.N; i++ {
 				haveAllNodes(test.nodeNames, test.nodes)
@@ -270,10 +271,6 @@ func TestHaveAllNodes(t *testing.T) {
 	}
 	// Test for the time complexity is O(n*n)
 	for _, test := range tests {
-		for i := 0; i < 40; i++ {
-			test.nodeNames = append(test.nodeNames, "worker-"+strconv.Itoa(i))
-			test.nodes = append(test.nodes, &st.MakeNode().Name("worker-"+strconv.Itoa(i)).Node)
-		}
 		t.Run(test.name, func(t *testing.T) {
 			if test.has != haveAllNodes(test.nodeNames, test.nodes) {
 				t.Errorf("test haveAllNodes wrong, test name is: %s", test.name)
@@ -282,7 +279,7 @@ func TestHaveAllNodes(t *testing.T) {
 	}
 	// Test for the time complexity is O(n)
 	for _, test := range tests {
-		for i := 40; i < 100; i++ {
+		for i := 0; i < haveAllNodesThreshold; i++ {
 			test.nodeNames = append(test.nodeNames, "worker-"+strconv.Itoa(i))
 			test.nodes = append(test.nodes, &st.MakeNode().Name("worker-"+strconv.Itoa(i)).Node)
 		}