@bart0sh 2023-03-03
I have different results for some reason:

goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/dynamicresources
cpu: Intel(R) Core(TM) i9-7920X CPU @ 2.90GHz

* original:
BenchmarkHaveAllNodes/have-nodes-10-24         	 6268983	       166.8 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-24         	  386126	      3100 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-24        	   96691	     **12394** ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-24        	    4819	    248655 ns/op	       0 B/op	       0 allocs/op
* PR:
BenchmarkHaveAllNodes/have-nodes-10-24         	 5643423	       188.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-24         	  345894	      3462 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-24        	   62052	     **19035** ns/op	    2888 B/op	       3 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-24        	   12471	     95487 ns/op	   20552 B/op	       3 allocs/op

BTW, we should also take into account that [podScheduling.Spec.PotentialNodes size is 128](https://github.com/kubernetes/kubernetes/blob/v1.26.2/staging/src/k8s.io/api/resource/v1alpha1/types.go#L214) atm

@sourcelliu @pohly WDYT?