@sourcelliu 2023-03-04
> I have different results for some reason
@bart0sh 
I have verified on both linux and darwin (and verified multiple versions of golang1.19, 1.20), and the test results are consistent with my previous test results. The test results of the two of us are inconsistent. This is really strange. I have not found any reason for the time being. The environmental information:
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/dynamicresources
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz

goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/dynamicresources
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
```