@bart0sh 2023-03-05
@sourcelliu I don't know why it happens, but it's a bit disturbing and needs to be investigated further in my opinion.

BTW,  using binary search(`sort.SearchStrings`) on longer slices should be quite fast. It requires sorted slice, but it worth trying if sorting overhead is compensated by a `SearchStrings` performance.