@sourcelliu 2023-03-05
> BTW, using binary search(`sort.SearchStrings`) on longer slices should be quite fast. It requires sorted slice, but it worth trying if sorting overhead is compensated by a `SearchStrings` performance.

That's a good idea. let me try.