@sourcelliu 2023-03-05
> BTW, using binary search(sort.SearchStrings) on longer slices should be quite fast. It requires sorted slice, but it worth trying if sorting overhead is compensated by a SearchStrings performance.

@bart0sh @pohly  I think we can consider only using binary search, the logic is simple, and in most cases (within 500 nodes), its effect is the best :
- original:
```
BenchmarkHaveAllNodes/have-nodes-10-16         	 7274658	       165.5 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-10-16         	 7138023	       167.0 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-10-16         	 7195632	       167.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  349687	      3529 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  330728	      3597 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  349957	      3576 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	   85381	     14078 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	   85586	     14240 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	   82430	     14248 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	    3987	    282241 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	    4184	    284312 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	    4318	    287753 ns/op	       0 B/op	       0 allocs/op
```
- PR(binary search):
```
BenchmarkHaveAllNodes/have-nodes-10-16         	 3643830	       320.2 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-10-16         	 3734691	       320.2 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-10-16         	 3715198	       322.9 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  604269	      1988 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  600572	      1987 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  578487	      1969 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	  266497	      4413 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	  258400	      4448 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	  260713	      4609 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	   30091	     39458 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	   30350	     39714 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-500-16        	   30136	     39749 ns/op	      24 B/op	       1 allocs/op
```

BTW, we should also take into account that [podScheduling.Spec.PotentialNodes size is 128](https://github.com/kubernetes/kubernetes/blob/v1.26.2/staging/src/k8s.io/api/resource/v1alpha1/types.go#L214) atm