@sourcelliu 2023-03-06
> It's still 2 times slower on 10 nodes and I'm not sure it's faster on 20 and 30 etc. Can you modify benchmark to be more granular, e.g. this way:

Done. the test result:
- original:
```
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/dynamicresources
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkHaveAllNodes/have-nodes-10-16         	 6482907	       171.1 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-20-16         	 2394909	       509.6 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-30-16         	 1000000	      1123 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-40-16         	  568514	      2126 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  362844	      3329 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-60-16         	  243204	      4892 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-70-16         	  173686	      6591 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-80-16         	  137946	      8553 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-90-16         	  110449	     10818 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	   90747	     13249 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-110-16        	   83829	     14268 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-120-16        	   77758	     15313 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-130-16        	   70564	     16791 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-140-16        	   63217	     18452 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-150-16        	   56512	     21557 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-160-16        	   51241	     23335 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-170-16        	   45817	     25903 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-180-16        	   41553	     28619 ns/op	       0 B/op	       0 allocs/op
BenchmarkHaveAllNodes/have-nodes-190-16        	   37660	     31797 ns/op	       0 B/op	       0 allocs/op
```
- PR:
```
goos: darwin
goarch: amd64
pkg: k8s.io/kubernetes/pkg/scheduler/framework/plugins/dynamicresources
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
BenchmarkHaveAllNodes/have-nodes-10-16         	 3271544	       343.1 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-20-16         	 1646317	       715.0 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-30-16         	 1000000	      1133 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-40-16         	  754983	      1632 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-50-16         	  579469	      2114 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-60-16         	  436404	      2662 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-70-16         	  391212	      3009 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-80-16         	  335790	      3590 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-90-16         	  295833	      4172 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-100-16        	  255979	      4748 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-110-16        	  225624	      5314 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-120-16        	  197174	      6089 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-130-16        	  162331	      6676 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-140-16        	  153862	      7750 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-150-16        	  143223	      8623 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-160-16        	  125427	      9424 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-170-16        	  115346	     10191 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-180-16        	  113673	     10510 ns/op	      24 B/op	       1 allocs/op
BenchmarkHaveAllNodes/have-nodes-190-16        	  100671	     11637 ns/op	      24 B/op	       1 allocs/op
```