@bart0sh 2023-03-06
> Done. the test result, We can find that starting from 30 nodes, the performance of binary search starts to outperform the original search algorithm:

Looks good, thanks. Why starting from 30? As far as I can see it still underperforms original code for 30 nodes.

It shows similar results on my linux and mac machines. Binary search starts to outperform original code at 35 nodes on both platforms:

* original:
  * linux:
  ```
  BenchmarkHaveAllNodes/have-nodes-10-24           6060434               171.1 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-15-24           4471293               250.2 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-20-24           2485680               471.3 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-25-24           1675332               727.4 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-30-24           1000000              1082 ns/op               0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-35-24            702949              1629 ns/op               0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-40-24            565663              2083 ns/op               0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-45-24            429234              2683 ns/op               0 B/op          0 allocs/op
  ```
  * mac:
  ```
  BenchmarkHaveAllNodes/have-nodes-10-8           11029656               103.1 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-15-8            6224546               191.6 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-20-8            3656398               324.5 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-25-8            2259595               532.2 ns/op             0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-30-8            1458933               794.8 ns/op             0 B/op          0 allocs/op 
  BenchmarkHaveAllNodes/have-nodes-35-8             978459              1114 ns/op               0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-40-8             768554              1501 ns/op               0 B/op          0 allocs/op
  BenchmarkHaveAllNodes/have-nodes-45-8             608484              1940 ns/op               0 B/op          0 allocs/op
  ```
* PR
  * linux:
  ```
  BenchmarkHaveAllNodes/have-nodes-10-24           3065745               416.6 ns/op            24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-15-24           1773772               577.6 ns/op            24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-20-24           1416486               843.9 ns/op            24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-25-24           1032832              1096 ns/op              24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-30-24            924386              1341 ns/op              24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-35-24            791736              1554 ns/op              24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-40-24            659244              1893 ns/op              24 B/op          1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-45-24            612360              2091 ns/op              24 B/op          1 allocs/op 
  ```
  * mac:
  ```
  BenchmarkHaveAllNodes/have-nodes-10-8         	 4685102	       248.9 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-15-8         	 3288391	       364.8 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-20-8         	 2023860	       589.8 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-25-8         	 1697974	       708.0 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-30-8         	 1356826	       883.4 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-35-8         	 1000000	      1081 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-40-8         	  881029	      1344 ns/op	      24 B/op	       1 allocs/op
  BenchmarkHaveAllNodes/have-nodes-45-8         	  799220	      1511 ns/op	      24 B/op	       1 allocs/op
  ```